+++ lib/ecc/Cargo.toml
@@ -7,3 +7,7 @@ description = "axVM elliptic curve cryptography library."
 
 [dependencies]
 axvm = { workspace = true }
+
+[features]
+default = []
+std = ["axvm/std"]
